+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Display multiple icons for mods that have local and workshop versions
 * Add button to open logs folder
 * Add ability to edit profile names
+* Add requiredModId to SpireEnum
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)